Rename schema list
