+++ assets/js/apps/search/components/filters/type-schema-list/TypeSchemaList.css
